update login and registartion update validation chages
