--- src/components/molecules/EmailField/index.tsx
@@ -1,32 +1,0 @@
-import Input from '@components/atoms/Input';
-import React from 'react';
-import { TextStyle, ViewStyle } from 'react-native';
-
-interface EmailFieldProps {
-  value: string;
-  onChangeText: (text: string) => void;
-  error?: string;
-  style?: ViewStyle | TextStyle;
-}
-
-const EmailField: React.FC<EmailFieldProps> = ({
-  value,
-  onChangeText,
-  error,
-  style,
-}) => {
-  return (
-    <Input
-      style={style}
-      placeholder="Email"
-      maxlength={25}
-      value={value}
-      onChangeText={onChangeText}
-      keyboardType='email-address'
-      autoCapitalize="none"
-      error={error}
-    />
-  );
-};
-
-export default EmailField;
